mlx5: fix scattered RX mode by avoiding completion queue resize
Resizing RX CQs when switching from non-SG to SG mode does not work
correctly with WQs. Since CQ size does not have to be the same as the number
of elements in the RX queue, always use the largest possible size.

This also applies to the maximum number of outstanding WRs in a WQ
(max_recv_wr).

Signed-off-by: Or Ami <ora@mellanox.com>
Signed-off-by: Adrien Mazarguil <adrien.mazarguil@6wind.com>
Signed-off-by: Nelio Laranjeiro <nelio.laranjeiro@6wind.com>