mlx5: fix queues reconfiguration support
Checking for RX queues while reconfiguring was necessary with the old RSS
API that required RX QPs to depend on a parent QP which had to be recreated.
It is not the case anymore since mlx5 uses hash RX queues.

Removing this check allows applications to update the number of queues.

Fixes: 5415b20e7ae3 ("mlx5: refactor RX code to use WQs instead of QPs for the new Verbs RSS API")

Signed-off-by: Nelio Laranjeiro <nelio.laranjeiro@6wind.com>