mlx4: ensure number of Rx queues is a power of 2
In the documentation it is specified that the hardware only supports a
number of RX queues if it is a power of 2.

Since ibv_exp_create_qp may not return an error when the number of
queues is unsupported by hardware, sanitize the value in dev_configure.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>
(cherry picked from commit bda501616c2f2f16c9d26bec472bf82d0f9ce6b7)
Signed-off-by: Adrien Mazarguil <adrien.mazarguil@6wind.com>