mlx5: handle CQE compression
Mini CQEs are used by the NIC when the PCI bandwidth is limited, in this case,
the first CQE64 received contains almost everything and the number following
CQE8, following CQEs are empty and used by the software for decompression.

From this:

      0           1          2           6         7         8
  +-------+  +---------+ +-------+   +-------+ +-------+ +-------+
  | CQE64 |  |  CQE64  | | CQE64 |   | CQE64 | | CQE64 | | CQE64 |
  |-------|  |---------| |-------|   |-------| |-------| |-------|
  | ..... |  | cqe8[0] | |       | . |       | |       | | ..... |
  | ..... |  | cqe8[1] | |       | . |       | |       | | ..... |
  | ..... |  | ....... | |       | . |       | |       | | ..... |
  | ..... |  | cqe8[7] | |       |   |       | |       | | ..... |
  +-------+  +---------+ +-------+   +-------+ +-------+ +-------+

We get this after decompression:

      0          1     ...     8
  +-------+  +-------+     +-------+
  | CQE64 |  | CQE64 |     | CQE64 |
  |-------|  |-------|     |-------|
  | ..... |  | ..... |  .  | ..... |
  | ..... |  | ..... |  .  | ..... |
  | ..... |  | ..... |  .  | ..... |
  | ..... |  | ..... |     | ..... |
  +-------+  +-------+     +-------+

This decompression cost a lot in resources knowing that the software only
needs the first CQE64 and the arrays of CQE8.

This patch provide avoid to recreate each cqe, it will ends by consuming more
CQEs than the uncompressed version because until all the compressed data is
not consumed, the CQEs associated will be kept.

Signed-off-by: Nelio Laranjeiro <nelio.laranjeiro@6wind.com>