ethdev: add speed capabilities
The speed capabilities of a device can be retrieved with
rte_eth_dev_info_get().

The new field speed_capa is initialized in the drivers without
taking care of device characteristics in this patch.
When the capabilities of a driver are accurate, the table in
overview.rst must be filled.

Signed-off-by: Marc Sune <marcdevel@gmail.com>
(cherry picked from commit e274f57322253a0da4b565f8df0175503c67258d)
Signed-off-by: Adrien Mazarguil <adrien.mazarguil@6wind.com>

Conflicts:
	doc/guides/nics/overview.rst
	doc/guides/rel_notes/release_16_04.rst
	drivers/net/nfp/nfp_net.c